+++ spec/support/cluster_creation_helpers.rb
@@ -9,7 +9,9 @@ module ClusterCreationHelpers
       subnet_azs: ['us-east-1a','us-east-1d','us-east-1f','us-east-1b'],
       include_analytics: true,
       cookbook_source_type: 's3',
-      include_utility: true
+      include_utility: true,
+      base_public_ami_id: "ami-12345",
+      base_private_ami_id: "ami-09876",
     }
   end
 end