+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * MI-97: RDS Aurora clusters
 * remove enhanced networking setup recipe from runlists
 * add mh-opsworks version compatability check
+* MI-134: find and use latest amis in cluster config
 
 ## 1.17.2 - 10/25/2018